Storybook: move Avatar data to the static folder
