+++ static/data/avatar.js
@@ -1,0 +1,21 @@
+import Image1 from '../avatars/1.png';
+import Image2 from '../avatars/2.png';
+import Image3 from '../avatars/3.png';
+import Image4 from '../avatars/4.png';
+import Image5 from '../avatars/5.png';
+import Image6 from '../avatars/6.png';
+import Image7 from '../avatars/7.png';
+import Image8 from '../avatars/8.png';
+
+const avatars = [
+  { image: Image1, count: 21, color: 'ruby', maxCount: 20 },
+  { image: Image2, count: 15, color: 'neutral', inactive: true },
+  { image: Image3, count: 5, color: 'neutral' },
+  { image: Image4, count: 3, color: 'ruby' },
+  { image: Image5, count: 0, color: 'ruby' },
+  { image: Image6, count: 4, color: 'ruby' },
+  { image: Image7, count: 5, color: 'neutral', inactive: true },
+  { image: Image8, count: 2, color: 'neutral', inactive: true },
+];
+
+export default avatars;
